Move BungieImageAndAmmo to loadout builder directory
