+++ src/app/item-popup/ItemPopupBody.scss
@@ -108,29 +108,6 @@
   padding: 2px 8px;
 }
 
-.ammo-type {
-  background-size: 26px;
-  height: 15px;
-  width: 20px;
-  margin-right: 4px;
-  background-color: rgba(0, 0, 0, 0.7);
-  background-position: center;
-  background-repeat: no-repeat;
-  border-radius: 5px;
-  margin-top: 1px;
-  flex-shrink: 0;
-}
-
-.ammo-primary {
-  background-image: url(https://www.bungie.net/common/destiny2_content/icons/dc4bb9bcdd4ae8a83fb9007a51d7d711.png);
-}
-.ammo-special {
-  background-image: url(https://www.bungie.net/common/destiny2_content/icons/b6d3805ca8400272b7ee7935b0b75c79.png);
-}
-.ammo-heavy {
-  background-image: url(https://www.bungie.net/common/destiny2_content/icons/9fa60d5a99c9ff9cea0fb6dd690f26ec.png);
-}
-
 .sheet {
   .move-popup {
     width: auto;
