+++ src/app/loadout-builder/BungieImageAndAmmo.m.scss
@@ -1,0 +1,32 @@
+.container {
+  position: relative;
+  height: min-content;
+  width: min-content;
+}
+
+.ammo {
+  background-size: 26px;
+  height: 15px;
+  width: 20px;
+  margin-right: 4px;
+  background-color: rgba(0, 0, 0, 0.7);
+  background-position: center;
+  background-repeat: no-repeat;
+  border-radius: 5px;
+  margin-top: 1px;
+  flex-shrink: 0;
+  position: absolute;
+  right: -4px;
+  bottom: -4px;
+  margin: 0 !important;
+}
+
+.ammoPrimary {
+  background-image: url(https://www.bungie.net/common/destiny2_content/icons/dc4bb9bcdd4ae8a83fb9007a51d7d711.png);
+}
+.ammoSpecial {
+  background-image: url(https://www.bungie.net/common/destiny2_content/icons/b6d3805ca8400272b7ee7935b0b75c79.png);
+}
+.ammoHeavy {
+  background-image: url(https://www.bungie.net/common/destiny2_content/icons/9fa60d5a99c9ff9cea0fb6dd690f26ec.png);
+}
