+++ src/app/loadout-builder/filter/LockedItem.tsx
@@ -1,7 +1,7 @@
-import BungieImageAndAmmo from 'app/dim-ui/BungieImageAndAmmo';
 import ConnectedInventoryItem from 'app/inventory/ConnectedInventoryItem';
 import DraggableInventoryItem from 'app/inventory/DraggableInventoryItem';
 import ItemPopupTrigger from 'app/inventory/ItemPopupTrigger';
+import BungieImageAndAmmo from 'app/loadout-builder/BungieImageAndAmmo';
 import React from 'react';
 import ArmorBucketIcon from '../ArmorBucketIcon';
 import ClosableContainer from '../ClosableContainer';
