+++ src/app/loadout-builder/locked-armor/SelectableBungieImage.tsx
@@ -4,10 +4,10 @@ import { InventoryBucket } from 'app/inventory/inventory-buckets';
 import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import { StatValue } from 'app/item-popup/PlugTooltip';
 import { SocketDetailsMod } from 'app/item-popup/SocketDetails';
+import BungieImageAndAmmo from 'app/loadout-builder/BungieImageAndAmmo';
 import { armorStatHashes } from 'app/search/search-filter-values';
 import clsx from 'clsx';
 import React from 'react';
-import BungieImageAndAmmo from '../../dim-ui/BungieImageAndAmmo';
 import ClosableContainer from '../ClosableContainer';
 import { LockedArmor2Mod, LockedItemType } from '../types';
 import styles from './SelectableBungieImage.m.scss';
